% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iedma.R
\name{iedma}
\alias{iedma}
\title{Prepare the matrix of euclidean distances from the original landmarks
coordinates.}
\usage{
iedma(x, p, k = 2)
}
\arguments{
\item{x}{a matrix containing the initial data of landmarks coordinates.}

\item{p}{number of landmarks.}

\item{k}{number of dimensions of the coordinates.}
}
\value{
A matrix with the lower triangular matrix of euclidean distance
  between all ladmarks coordinates for each individual placed in a single row.
}
\description{
This function assists in the process of importing a file with original
landmarks coordinates for each studied individuals and transforming these
data into the lower triangular matrix of euclidean distance between all
ladmarks coordinates for each individual placed in a single row.
}
\examples{
new_file <- iedma(original_file, 10)

}
